Move SensorLogReader to separate file
